Added Daily Averages, added more details for items, fixed keyboard in settings
Settings is not in a good state, the keyboard is fixed but now there is a problem where you can pull the screen and see white instead of black, also height should have feet and inches. The save button is not implemented, and the styling generally doesn't look very good. Need to fix in future.